requirements.txt
azure-functions==1.19.0
openai==1.30.1
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchain-experimental==0.0.59
langchain-openai==0.1.7
pydantic==2.7.1
pydantic-core==2.18.2
numpy==1.26.4
azure-identity
requests>=2.25.1